rename jsx to tsx, fix imports
