--- packages/client/src/LoginPage/GoogleAuth.jsx
@@ -1,5 +1,0 @@
-export default class GoogleAuth extends React.Component {
-  render() {
-    return <div className="py-3 border">GoogleAuth PlaceHolder</div>;
-  }
-}
