+++ packages/client/src/LoginPage/GoogleAuth.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+export default class GoogleAuth extends React.Component {
+    render() {
+        return <div className="py-3 border">GoogleAuth PlaceHolder</div>
+    }
+}
